requirements.txt
pandas==1.3.5
Django==3.2.13
djangorestframework==3.13.1
markdown==3.3.7
django-filter==21.1
django-extensions==3.1.5
